daily commit
Change everything graph and table to flat style.
Create tooltips for every graph
Finish the links in the dashboard page
Put data into the Table of performance page
When you select one day in the champagne, it show the column chart
instead of line chart

Bugs fixed:
Axis Labels in the graph. Now it shows more labels
Table can’t be updated after clicking valider button.
When the cursor moves to a cell in the dashboard page, the mouse doesn’t
change. Now it will change to a pointer.